% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enrich.internal.R
\name{ALLEXTID.DOLite}
\alias{ALLEXTID.DOLite}
\title{ALLEXTID.DOLite}
\usage{
\method{ALLEXTID}{DOLite}(organism, use_internal_data)
}
\arguments{
\item{organism}{organism}

\item{use_internal_data}{logical}
}
\description{
ALLEXTID.DOLite
}

