\docType{methods}
\name{cnetplot}
\alias{cnetplot}
\alias{cnetplot,enrichResult,ANY-method}
\title{cnetplot method}
\arguments{
  \item{...}{Additional argument list}

  \item{showCategory}{number of category plotted}

  \item{categorySize}{one of geneNum or pvalue}

  \item{logFC}{fold change of expression value}

  \item{output}{one of fixed or interactive}
}
\value{
  plot

  plot
}
\description{
  cnetplot method generics

  cnetplot method generics
}
\author{
  Guangchuang Yu \url{http://ygc.name}

  Guangchuang Yu \url{http://ygc.name}
}

