% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cnetplot.R
\name{cnetplot_internal}
\alias{cnetplot_internal}
\title{cnetplot_internal}
\usage{
cnetplot_internal(inputList, categorySize = "geneNum", showCategory = 5,
  pvalue = NULL, foldChange = NULL, fixed = TRUE, DE.foldChange = NULL,
  ...)
}
\arguments{
\item{inputList}{a list of gene IDs}

\item{categorySize}{setting category size}

\item{showCategory}{number of categories to plot}

\item{pvalue}{pvalue}

\item{foldChange}{fold Change}

\item{fixed}{logical}

\item{DE.foldChange}{logical}

\item{...}{additional parameters}
}
\value{
plotted igraph object.
}
\description{
plot function of gene Concept Net.
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

