% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{compareClusterResult-class}
\alias{compareClusterResult-class}
\alias{show,compareClusterResult-method}
\alias{summary,compareClusterResult-method}
\alias{plot,compareClusterResult-method}
\title{Class "compareClusterResult"
This class represents the comparison result of gene clusters by GO
categories at specific level or GO enrichment analysis.}
\description{
Class "compareClusterResult"
This class represents the comparison result of gene clusters by GO
categories at specific level or GO enrichment analysis.
}
\section{Slots}{

\describe{
\item{\code{compareClusterResult}}{cluster comparing result}

\item{\code{geneClusters}}{a list of genes}

\item{\code{fun}}{one of groupGO, enrichGO and enrichKEGG}

\item{\code{gene2Symbol}}{gene ID to Symbol}

\item{\code{keytype}}{Gene ID type}

\item{\code{readable}}{logical flag of gene ID in symbol or not.}

\item{\code{.call}}{function call}

\item{\code{termsim}}{Similarity between term}

\item{\code{method}}{method of calculating the similarity between nodes}

\item{\code{dr}}{dimension reduction result}
}}

\seealso{
\code{\linkS4class{enrichResult}}
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}
\keyword{classes}
