% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DOSE-package.R
\docType{data}
\name{DataSet}
\alias{DO2ALLEG}
\alias{DO2EG}
\alias{DOSEEnv}
\alias{DataSet}
\alias{EG2ALLDO}
\alias{EG2DO}
\alias{IC}
\alias{Info_Contents_human_DO}
\alias{geneList}
\title{Datasets}
\description{
Information content and DO term to entrez gene IDs mapping
}
\keyword{datasets}

