% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose.R
\name{dose}
\alias{dose}
\title{The DOSE index in COPD}
\usage{
dose(mmrc, fev, smoking_status, exacerbation)
}
\arguments{
\item{mmrc}{A number for the dyspnea scale of modified Medical Research Council (mMRC)}

\item{fev}{A number for forced expiratory volume in one second (FEV1) expressed as percentage}

\item{smoking_status}{Binary variable of patient's smoking status (yes as 1 or no as 0)}

\item{exacerbation}{The number of exacerbations of COPD during the previous year}
}
\value{
The DOSE index
}
\description{
The DOSE index in COPD
}
\examples{
dose(mmrc=0, fev=55, smoking_status=1, exacerbation=2)
}
