% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpretDisease.R
\name{show,doseInterpretResult-method}
\alias{show,doseInterpretResult-method}
\title{Show a canonical interpretation result}
\usage{
\S4method{show}{doseInterpretResult}(object)
}
\arguments{
\item{object}{A `doseInterpretResult` object.}
}
\value{
The object is shown to the console and returned invisibly.
}
\description{
Show a canonical interpretation result
}
