% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doSim.R
\name{doseSim}
\alias{doseSim}
\title{doSim}
\usage{
doseSim(DOID1, DOID2, measure = "Wang", ont = "DO")
}
\arguments{
\item{DOID1}{DO term, MPO term or HPO term vector}

\item{DOID2}{DO term, MPO term or HPO term vector}

\item{measure}{one of "Wang", "Resnik", "Rel", "Jiang", "Lin", and "TCSS".}

\item{ont}{one of "DO" and "MPO"}
}
\value{
score matrix
}
\description{
measuring similarities between two DO term vectors.
}
\details{
provide two term vectors, this function will calculate their similarities.
}
