\docType{class}
\name{enrichDOResult-class}
\alias{enrichDOResult-class}
\alias{show,enrichDOResult-method}
\alias{summary,enrichDOResult-method}
\title{Class "enrichDOResult"
This class represents the result of DO enrichment analysis.}
\description{
  Class "enrichDOResult" This class represents the result
  of DO enrichment analysis.
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}
\seealso{
  \code{\link{enrichDO}}
}
\keyword{classes}

