% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichMap.R
\name{enrichMap}
\alias{enrichMap}
\title{enrichMap}
\usage{
enrichMap(x, n = 50, ...)
}
\arguments{
\item{x}{gseaResult or enrichResult object}

\item{n}{maximum number of category to shown}

\item{...}{additional parameter}
}
\value{
figure
}
\description{
enrichment map
}
\details{
enrichment map
}
\author{
G Yu
}
