\name{EXTID2NAME}
\alias{EXTID2NAME}
\title{convert gene IDs to gene Names}
\usage{
  EXTID2NAME(geneID, organism)
}
\arguments{
  \item{geneID}{a vector of gene IDs}

  \item{organism}{only "human" supported}
}
\value{
  a vector of gene names.
}
\description{
  convert gene IDs to gene Names.
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}

