% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{EXTID2TERMID}
\alias{EXTID2TERMID}
\title{EXTID2TERMID}
\usage{
EXTID2TERMID(gene, organism, use.KEGG.db)
}
\arguments{
\item{gene}{gene ID vector}

\item{organism}{organism}

\item{use.KEGG.db}{logical}
}
\description{
Mapping External ID to Ontology Term ID
}

