% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.gseaResult}
\alias{fortify.gseaResult}
\title{fortify.gseaResult}
\usage{
\method{fortify}{gseaResult}(model, data, geneSetID, ...)
}
\arguments{
\item{model}{gseaResult object}

\item{data}{not used.}

\item{geneSetID}{gene set ID}

\item{...}{additional parameter}
}
\value{
figure
}
\description{
fortify.gseaResult
}
\author{
G Yu
}

