% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gene2DO}
\alias{gene2DO}
\title{convert Gene ID to DO Terms}
\usage{
gene2DO(gene, organism = NULL, ont = "HDO")
}
\arguments{
\item{gene}{entrez gene ID}

\item{organism}{species of the Entrez gene ID. If omitted, it is inferred
from `ont`.}

\item{ont}{ont}
}
\value{
DO Terms
}
\description{
provide gene ID, this function will convert to the corresponding DO Terms
}
\author{
Guangchuang Yu \url{https://yulab-smu.top}
}
