% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{getALLEG}
\alias{getALLEG}
\title{getALLEG}
\usage{
getALLEG(organism)
}
\arguments{
\item{organism}{species}
}
\value{
entrez gene ID vector
}
\description{
get all entrezgene ID of a specific organism
}
\author{
Yu Guangchuang
}

