% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gseAnalyzer.R
\name{getGeneSet.DO}
\alias{getGeneSet.DO}
\title{getGeneSet.DO}
\usage{
\method{getGeneSet}{DO}(setType = "DO", organism, use_internal_data = TRUE)
}
\arguments{
\item{setType}{setType}

\item{organism}{organism}

\item{use_internal_data}{logical}
}
\description{
getGeneSet.DO
}

