% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gseAnalyzer.R
\name{getGeneSet.DOLite}
\alias{getGeneSet.DOLite}
\title{getGeneSet.DOLite}
\usage{
\method{getGeneSet}{DOLite}(setType = "DOLite", organism)
}
\arguments{
\item{setType}{setType}

\item{organism}{organism}
}
\description{
getGeneSet.DOLite
}

