% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getGeneSet}
\alias{getGeneSet}
\title{getGeneSet}
\usage{
getGeneSet(setType, organism, ...)
}
\arguments{
\item{setType}{type of gene sets}

\item{organism}{organism}

\item{...}{additional parameter}
}
\description{
preparing geneSets for gene set enrichment analysis
}

