% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{GSEA_internal}
\alias{GSEA_internal}
\title{GSEA_internal}
\usage{
GSEA_internal(
  geneList,
  exponent,
  minGSSize,
  maxGSSize,
  eps,
  pvalueCutoff,
  pAdjustMethod,
  verbose,
  seed = FALSE,
  USER_DATA,
  by = "fgsea",
  ...
)
}
\arguments{
\item{geneList}{order ranked geneList}

\item{exponent}{weight of each step}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of each geneSet for analyzing}

\item{eps}{This parameter sets the boundary for calculating the p value.}

\item{pvalueCutoff}{p value Cutoff}

\item{pAdjustMethod}{p value adjustment method}

\item{verbose}{print message or not}

\item{seed}{set seed inside the function to make result reproducible. FALSE by default.}

\item{USER_DATA}{annotation data}

\item{by}{one of 'fgsea' or 'DOSE'}

\item{...}{other parameter}
}
\value{
gseaResult object
}
\description{
generic function for gene set enrichment analysis
}
\author{
Yu Guangchuang
}
