% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gseAnalyzer}
\alias{gseAnalyzer}
\title{Gene Set Enrichment Analysis}
\usage{
gseAnalyzer(geneList, setType, organism = "human", exponent = 1,
  nPerm = 1000, minGSSize = 10, pvalueCutoff = 0.05,
  pAdjustMethod = "BH", verbose = TRUE)
}
\arguments{
\item{geneList}{order ranked geneList}

\item{setType}{Type of geneSet}

\item{organism}{organism}

\item{exponent}{weight of each step}

\item{nPerm}{permutation numbers}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{pvalueCutoff}{pvalue Cutoff}

\item{pAdjustMethod}{p value adjustment method}

\item{verbose}{print message or not}
}
\value{
gseaResult object
}
\description{
perform gsea analysis
}
\author{
Yu Guangchuang
}
\keyword{manip}

