\docType{class}
\name{gseaResult-class}
\alias{gseaResult-class}
\alias{gseahResult-class}
\alias{plot,gseaResult-method}
\alias{show,gseaResult-method}
\alias{summary,gseaResult-method}
\title{Class "gseaResult"
This class represents the result of GSEA analysis}
\description{
  Class "gseaResult" This class represents the result of
  GSEA analysis
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}
\seealso{
  \code{\link{gseaplot}}
}
\keyword{classes}

