% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseAnalyzer.R
\name{gseDGN}
\alias{gseDGN}
\title{DisGeNET Gene Set Enrichment Analysis}
\usage{
gseDGN(
  geneList,
  exponent = 1,
  nPerm = 1000,
  minGSSize = 10,
  maxGSSize = 500,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  verbose = TRUE,
  adaptive = FALSE,
  minPerm = 1000,
  maxPerm = 10000,
  ...
)
}
\arguments{
\item{geneList}{order ranked geneList}

\item{exponent}{weight of each step}

\item{nPerm}{permutation numbers}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of each geneSet for analyzing}

\item{pvalueCutoff}{pvalue Cutoff}

\item{pAdjustMethod}{p value adjustment method}

\item{verbose}{print message or not}

\item{adaptive}{logical, use adaptive permutation or not (default: FALSE)}

\item{minPerm}{minimum number of permutations for adaptive mode (default: 1000)}

\item{maxPerm}{maximum number of permutations for adaptive mode (default: 10000)}

\item{...}{other parameter}
}
\value{
gseaResult object
}
\description{
perform gsea analysis
}
\author{
Guangchuang Yu
}
\keyword{manip}
