% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsfilter.R
\name{gsfilter}
\alias{gsfilter}
\title{gsfilter}
\usage{
gsfilter(x, by = "GSSize", min = NA, max = NA)
}
\arguments{
\item{x}{instance of enrichResult or compareClusterResult}

\item{by}{one of 'GSSize' or 'Count'}

\item{min}{minimal size}

\item{max}{maximal size}
}
\value{
update object
}
\description{
filter enriched result by gene set size or gene count
}
\author{
Guangchuang Yu
}

