% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doSim.R
\name{hpoSim}
\alias{hpoSim}
\title{doSim}
\usage{
hpoSim(DOID1, DOID2, measure = "Wang")
}
\arguments{
\item{DOID1}{HPO term vector}

\item{DOID2}{HPO term vector}

\item{measure}{one of "Wang", "Resnik", "Rel", "Jiang", "Lin", and "TCSS".}
}
\value{
score matrix
}
\description{
measuring similarities between two MPO term vectors.
}
\details{
provide two HPO term vectors, this function will calculate their similarities.
}
