% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joyplot.R
\name{joyplot}
\alias{joyplot}
\title{joyplot}
\usage{
joyplot(x, showCategory = 30, fill = "p.adjust", core_enrichment = TRUE)
}
\arguments{
\item{x}{gseaResult object}

\item{showCategory}{number of categories for plotting}

\item{fill}{one of "pvalue", "p.adjust", "qvalue"}

\item{core_enrichment}{whether only using core_enriched genes}
}
\value{
ggplot object
}
\description{
joyplot for GSEA result
}
\author{
guangchuang yu
}

