% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{load_OrgDb}
\alias{load_OrgDb}
\title{load_OrgDb}
\usage{
load_OrgDb(OrgDb)
}
\arguments{
\item{OrgDb}{OrgDb object or OrgDb name}
}
\value{
OrgDb object
}
\description{
load OrgDb
}
\author{
Guangchuang Yu
}

