% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doSim.R
\name{mpoSim}
\alias{mpoSim}
\title{doSim}
\usage{
mpoSim(DOID1, DOID2, measure = "Wang")
}
\arguments{
\item{DOID1}{MPO term vector}

\item{DOID2}{MPO term vector}

\item{measure}{one of "Wang", "Resnik", "Rel", "Jiang", "Lin", and "TCSS".}
}
\value{
score matrix
}
\description{
measuring similarities between two MPO term vectors.
}
\details{
provide two MPO term vectors, this function will calculate their similarities.
}
