% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ratio.R
\name{parse_ratio}
\alias{parse_ratio}
\title{parse_ratio}
\usage{
parse_ratio(ratio)
}
\arguments{
\item{ratio}{character vector of ratio to parse}
}
\value{
A numeric vector (double) of parsed ratio
}
\description{
parse character ratio to double value, such as 1/5 to 0.2
}
\author{
Guangchuang Yu
}
