% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouseModels.R
\name{rankMouseModels}
\alias{rankMouseModels}
\title{Rank mouse disease models from human query genes}
\usage{
rankMouseModels(
  gene,
  disease_model_file,
  homology_file,
  gene_pheno_file,
  top = 50
)
}
\arguments{
\item{gene}{A character vector of human Entrez gene IDs.}

\item{disease_model_file}{Path to an `MGI_DiseaseMouseModel.rpt`-format file.}

\item{homology_file}{Path to a `HOM_MouseHumanSequence.rpt`-format file.}

\item{gene_pheno_file}{Path to an `MGI_GenePheno.rpt`-format file.}

\item{top}{Maximum number of mouse-model targets to return.}
}
\value{
A `doseInterpretResult` with `target_type = "mouse_model"`.
}
\description{
Rank mouse disease models from human query genes
}
