% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{rebuildAnnoData}
\alias{rebuildAnnoData}
\title{rebuiding annotation data}
\usage{
rebuildAnnoData(file)
}
\arguments{
\item{file}{do_rif.human.txt}
}
\description{
rebuilding entrez and DO mapping datasets
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

