% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{facet_grid}
\alias{get_organism}
\alias{set_auto_update}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2:facet_grid]{facet_grid()}}}

  \item{GOSemSim}{\code{\link[GOSemSim:get_organism]{get_organism()}}, \code{\link[GOSemSim:set_auto_update]{set_auto_update()}}}
}}

