\docType{methods}
\name{setReadable}
\alias{setReadable}
\title{Methods mapping gene ID to gene symbol for \code{enrichResult} instance}
\arguments{
  \item{params}{A \code{enrichResult} instance.}
}
\value{
  A \code{enrichResult} instance.
}
\description{
  setReadable method for \code{enrichResult} instance
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}

