% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichResult.R
\name{setReadable}
\alias{setReadable}
\title{setReadable}
\usage{
setReadable(x, OrgDb)
}
\arguments{
\item{x}{enrichResult Object}

\item{OrgDb}{OrgDb}
}
\value{
enrichResult Object
}
\description{
mapping geneID to gene Symbol
}
\author{
Yu Guangchuang
}

