% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cnetplot.R
\name{setting.graph.attributes}
\alias{setting.graph.attributes}
\title{setting.graph.attributes}
\usage{
setting.graph.attributes(g, node.size = 8, node.color = "#B3B3B3",
  edege.width = 2, edege.color = "#8DA0CB")
}
\arguments{
\item{g}{igraph object}

\item{node.size}{size of node}

\item{node.color}{color of node}

\item{edege.width}{edege width}

\item{edege.color}{color of edege}
}
\value{
igraph object
}
\description{
setting basic attributes of a graph
}
\details{
setting size and color of node and edege
}
\author{
Yu Guangchuang
}

