\docType{methods}
\name{show}
\alias{show}
\title{show method}
\arguments{
  \item{object}{A \code{enrichResult} instance.}

  \item{object}{A \code{gseaResult} instance.}
}
\value{
  message

  message
}
\description{
  show method for \code{enrichResult} instance

  show method for \code{gseaResult} instance
}
\author{
  Guangchuang Yu \url{http://ygc.name}

  Guangchuang Yu \url{http://ygc.name}
}

