\docType{methods}
\name{sim}
\alias{sim}
\title{Methods for calculating semantic similarity}
\arguments{
  \item{params}{A \code{DOParams} instance.}
}
\value{
  Semantic similarity value or matrix.
}
\description{
  sim method for \code{DOParams} instance
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}

