% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichResult.R, R/gseAnalyzer.R
\docType{methods}
\name{[[,enrichResult-method}
\alias{[[,enrichResult-method}
\alias{[[,gseaResult-method}
\title{[[ method}
\usage{
\S4method{[[}{enrichResult}(x, term)

\S4method{[[}{gseaResult}(x, term)
}
\arguments{
\item{x}{enrichResult object}

\item{term}{term to query}
}
\description{
accessing gene set

accessing core enrichment genes
}

