% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichResult.R, R/gseaResult.R
\docType{methods}
\name{[[,enrichResult-method}
\alias{[[,enrichResult-method}
\alias{[[,gseaResult-method}
\title{[[ method}
\usage{
\S4method{[[}{enrichResult}(x, i)

\S4method{[[}{gseaResult}(x, i)
}
\arguments{
\item{x}{enrichResult object}

\item{i}{term to query}
}
\description{
accessing gene set

accessing core enrichment genes
}

