% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichResult.R, R/gseaResult.R
\docType{methods}
\name{$,enrichResult-method}
\alias{$,enrichResult-method}
\alias{$,gseaResult-method}
\title{$ method}
\usage{
\S4method{$}{enrichResult}(x, name)

\S4method{$}{gseaResult}(x, name)
}
\arguments{
\item{x}{enrichResult object}

\item{name}{string}
}
\description{
accessing enriched result
}

