% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{TERM2NAME}
\alias{TERM2NAME}
\title{TERM2NAME}
\usage{
TERM2NAME(term, organism, ...)
}
\arguments{
\item{term}{term ID vector}

\item{organism}{organism}

\item{...}{additional parameter}
}
\description{
Mapping Ontology Term ID to Name Symbol or Description
}

