% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetplot.R
\name{upsetplot}
\alias{upsetplot}
\title{upsetplot}
\usage{
upsetplot(x, n = 10)
}
\arguments{
\item{x}{enrichResult object}

\item{n}{number of categories}
}
\value{
plot
}
\description{
upsetplot
}
\examples{
\dontrun{
require(DOSE)
data(geneList)
de=names(geneList)[1:100]
x <- enrichDO(de)
upsetplot(x, 8)
}
}
\author{
Guangchuang Yu
}

