% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enrich.internal.R
\name{ALLEXTID.DO}
\alias{ALLEXTID.DO}
\title{ALLEXTID.DO}
\usage{
\method{ALLEXTID}{DO}(organism, use.KEGG.db)
}
\arguments{
\item{organism}{organism}

\item{use.KEGG.db}{logical}
}
\description{
ALLEXTID.DO
}

