% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ALLEXTID}
\alias{ALLEXTID}
\title{ALLEXTID}
\usage{
ALLEXTID(organism, use.KEGG.db)
}
\arguments{
\item{organism}{organism}

\item{use.KEGG.db}{logical}
}
\description{
Get all background External ID.
}

