% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ALLEXTID}
\alias{ALLEXTID}
\title{ALLEXTID}
\usage{
ALLEXTID(organism, ...)
}
\arguments{
\item{organism}{organism}

\item{...}{additional parameter}
}
\description{
Get all background External ID.
}

