% Generated by roxygen2 (4.0.2): do not edit by hand
\name{barplot.enrichResult}
\alias{barplot.enrichResult}
\title{barplot}
\usage{
\method{barplot}{enrichResult}(height, font.size = 12, title = "", ...)
}
\arguments{
\item{height}{enrichResult object}

\item{font.size}{font size}

\item{title}{plot title}

\item{...}{other parameter, ignored}
}
\description{
barplot
}

