% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/enrichDO.R
\docType{methods}
\name{cnetplot}
\alias{cnetplot}
\alias{cnetplot,enrichResult,ANY-method}
\title{cnetplot method}
\usage{
cnetplot(x, showCategory = 5, categorySize = "geneNum", foldChange = NULL,
  fixed = TRUE, ...)

cnetplot(x, showCategory=5, categorySize="geneNum", foldChange=NULL, fixed=TRUE, ...)
}
\arguments{
\item{x}{enrichResult object}

\item{showCategory}{number of category plotted}

\item{categorySize}{one of geneNum or pvalue}

\item{foldChange}{fold change of expression value}

\item{fixed}{logical}

\item{...}{additional parameter}
}
\value{
plot

plot
}
\description{
cnetplot method generics

cnetplot method generics
}
\author{
Guangchuang Yu \url{http://ygc.name}

Guangchuang Yu \url{http://ygc.name}
}

