% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{computeIC}
\alias{computeIC}
\title{compute information content}
\usage{
computeIC(ont = "DO")
}
\arguments{
\item{ont}{one of "DO" and "MPO"}
}
\description{
compute information content
}
\author{
Guangchuang Yu \url{http://guangchuangyu.github.io}
}
