\name{computeIC}
\alias{computeIC}
\title{compute information content}
\usage{
  computeIC(ont = "DO", organism = "human")
}
\arguments{
  \item{ont}{"DO"}

  \item{organism}{"human"}
}
\value{
  NULL
}
\description{
  compute information content
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}

