% Generated by roxygen2 (4.0.2): do not edit by hand
\name{computeIC}
\alias{computeIC}
\title{compute information content}
\usage{
computeIC(ont = "DO", organism = "human")
}
\arguments{
\item{ont}{"DO"}

\item{organism}{"human"}
}
\description{
compute information content
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

