% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{computeIC}
\alias{computeIC}
\title{compute information content}
\usage{
computeIC(ont = "DO", organism = "human")
}
\arguments{
\item{ont}{"DO"}

\item{organism}{"human"}
}
\description{
compute information content
}
\author{
Guangchuang Yu \url{http://guangchuangyu.github.io}
}
