\docType{data}
\name{DataSet}
\alias{AL1}
\alias{AL1expr}
\alias{DO2ALLEG}
\alias{DO2EG}
\alias{DOLite2EG}
\alias{DOLiteTerm}
\alias{DOSEEnv}
\alias{DataSet}
\alias{EG2ALLDO}
\alias{EG2DO}
\alias{EG2DOLite}
\alias{IC}
\alias{Info_Contents_human_DO}
\title{Datasets
Information content and DO term to entrez gene IDs mapping}
\description{
  Datasets Information content and DO term to entrez gene
  IDs mapping
}
\keyword{datasets}

