% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOSE-package.R
\docType{data}
\name{DataSet}
\alias{DataSet}
\alias{EG2DO}
\alias{DO2EG}
\alias{EG2ALLDO}
\alias{DO2ALLEG}
\alias{DOIC}
\alias{dotbl}
\alias{EG2DOLite}
\alias{DOLite2EG}
\alias{DOLiteTerm}
\alias{geneList}
\alias{NCG_EXTID2PATHID}
\alias{NCG_PATHID2EXTID}
\alias{NCG_PATHID2NAME}
\alias{DGN_EXTID2PATHID}
\alias{DGN_PATHID2EXTID}
\alias{DGN_PATHID2NAME}
\alias{VDGN_EXTID2PATHID}
\alias{VDGN_PATHID2EXTID}
\alias{VDGN_PATHID2NAME}
\title{Datasets}
\description{
Information content and DO term to entrez gene IDs mapping
}
\keyword{datasets}
