% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOSE-package.R
\docType{data}
\name{DataSet}
\alias{DO2ALLEG}
\alias{DO2EG}
\alias{DOIC}
\alias{DOLite2EG}
\alias{DOLiteTerm}
\alias{DOSEEnv}
\alias{DataSet}
\alias{EG2ALLDO}
\alias{EG2DO}
\alias{EG2DOLite}
\alias{NCG_DOSE_Env}
\alias{dotbl}
\alias{geneList}
\title{Datasets}
\description{
Information content and DO term to entrez gene IDs mapping
}
\keyword{datasets}

