\docType{class}
\name{DOParams-class}
\alias{DOParams-class}
\alias{sim,DOParams-method}
\title{Class "DOParams"
This class contains parameters for calculating DO semantic similarity among DO term or Gene list.}
\description{
  Class "DOParams" This class contains parameters for
  calculating DO semantic similarity among DO term or Gene
  list.
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}
\seealso{
  \code{\link{sim}}
}
\keyword{classes}

