% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOSE-package.R
\docType{package}
\name{DOSE-package}
\alias{DOSE}
\alias{DOSE-package}
\title{DOSE: Disease Ontology Semantic and Enrichment analysis}
\description{
Provides ontology-aware methods for disease and phenotype knowledge mining. DOSE supports semantic similarity analysis of disease and phenotype ontology terms, genes, and gene clusters using methods including Resnik, Schlicker, Jiang, Lin, and Wang. It also provides over-representation analysis and gene set enrichment analysis for interpreting gene vectors and ranked gene lists in disease, phenotype, and cancer contexts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yulab-smu.top/contribution-knowledge-mining/}
  \item Report bugs at \url{https://github.com/GuangchuangYu/DOSE/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com}

Authors:
\itemize{
  \item Guangchuang Yu \email{guangchuangyu@gmail.com}
}

Other contributors:
\itemize{
  \item Li-Gen Wang \email{reeganwang020@gmail.com} [contributor]
  \item Vladislav Petyuk \email{petyuk@gmail.com} [contributor]
  \item Giovanni Dall'Olio \email{giovanni.dallolio@upf.edu} [contributor]
}

}
\keyword{internal}
