% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOSE-package.R
\docType{package}
\name{DOSE-package}
\alias{DOSE}
\alias{DOSE-package}
\title{DOSE: Disease Ontology Semantic and Enrichment analysis}
\description{
This package implements five methods proposed by Resnik, Schlicker, Jiang, Lin and Wang respectively for measuring semantic similarities among DO terms and gene products. Enrichment analyses including hypergeometric model and gene set enrichment analysis are also implemented for discovering disease associations of high-throughput biological data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yulab-smu.top/biomedical-knowledge-mining-book/}
  \item Report bugs at \url{https://github.com/GuangchuangYu/DOSE/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com}

Other contributors:
\itemize{
  \item Li-Gen Wang \email{reeganwang020@gmail.com} [contributor]
  \item Vladislav Petyuk \email{petyuk@gmail.com} [contributor]
  \item Giovanni Dall'Olio \email{giovanni.dallolio@upf.edu} [contributor]
  \item Erqiang Hu \email{13766876214@163.com} [contributor]
}

}
\keyword{internal}
