\name{DOSE-package}
\alias{DOSE-package}
\alias{DOSE}
\docType{package}
\title{
	  Disease Ontology Semantic and Enrichment analysis
}
\description{
	Implemented five methods proposed by Resnik, Schlicker, Jiang, Lin and Wang respectively for measuring DO semantic similarities, and hypergeometric test for enrichment analysis.
}

\details{
	This package is designed to estimate DO-based semantic similarity measurement and enrichment analysis.

	\tabular{ll}{
		Package: \tab DOSE\cr
		Type: \tab Package\cr
		Version: \tab 0.99.4\cr
		Date: \tab 06-10-2011\cr
		biocViews:\tab Bioinformatics, Annotation\cr
		Depends:\tab \cr
		Imports: \tab methods, AnnotationDbi, DO.db\cr
		Suggests:\tab clusterProfiler, GOSemSim\cr
		License: \tab Artistic-2.0\cr
	}
}

\author{
	Guangchuang Yu

	Maintainer: Guangchuang Yu <guangchuangyu@gmail.com>
}

\seealso{
  \linkS4class{DOParams},
  \linkS4class{enrichDOResult}
}

\keyword{ package }
