% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DOSE-package.R
\docType{package}
\name{DOSE-package}
\alias{DOSE}
\alias{DOSE-package}
\title{Disease Ontology Semantic and Enrichment analysis
Implemented five methods proposed by Resnik, Schlicker, Jiang, Lin and Wang respectively for measuring DO semantic similarities, and hypergeometric test for enrichment analysis.}
\description{
This package is designed to estimate DO-based semantic similarity
measurement and enrichment analysis.
}
\details{
\tabular{ll}{ Package: \tab DOSE\cr Type: \tab Package\cr Version: \tab
2.3.5\cr Date: \tab 2-27-2012\cr biocViews:\tab Bioinformatics,
Annotation\cr Depends:\tab \cr Imports: \tab methods, AnnotationDbi,
DO.db\cr Suggests:\tab clusterProfiler, GOSemSim\cr License: \tab
Artistic-2.0\cr }
}
\author{
Guangchuang Yu, Li-Gen Wang

Maintainer: Guangchuang Yu \email{guangchuangyu@gmail.com}
}
\seealso{
\linkS4class{enrichResult}
}
\keyword{package}

