% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dose_params}
\alias{dose_params}
\title{Shared parameters for DOSE functions}
\arguments{
\item{gene}{a vector of entrez gene id}

\item{organism}{one of "hsa" and "mmu"}

\item{ont}{one of "HDO", "HPO" or "MPO"}

\item{pvalueCutoff}{pvalue cutoff}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{universe}{background genes}

\item{minGSSize}{minimal size of genes annotated by ontology term for testing}

\item{maxGSSize}{maximal size of each geneSet for analyzing}

\item{qvalueCutoff}{qvalue cutoff}

\item{readable}{whether mapping gene ID to gene Name}

\item{geneList}{order ranked geneList}

\item{exponent}{weight of each step}

\item{nPerm}{permutation numbers}

\item{verbose}{print message or not}

\item{adaptive}{logical, use adaptive permutation or not (default: FALSE)}

\item{minPerm}{minimum number of permutations for adaptive mode (default: 1000)}

\item{maxPerm}{maximum number of permutations for adaptive mode (default: 10000)}

\item{method}{method of GSEA, one of "multilevel", "permute", "sample"}
}
\description{
Shared parameters for DOSE functions
}
