% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/enrichResult.R
\docType{methods}
\name{dotplot}
\alias{dotplot}
\alias{dotplot,enrichResult,ANY-method}
\alias{dotplot,enrichResult-method}
\title{dotplot method}
\usage{
dotplot(object, ...)

\S4method{dotplot}{enrichResult}(object, x = "geneRatio",
  colorBy = "p.adjust", showCategory = 10, font.size = 12, title = "")
}
\arguments{
\item{object}{an instance of enrichResult}

\item{...}{additional parameter}

\item{x}{variable for x axis}

\item{colorBy}{one of 'pvalue', 'p.adjust' and 'qvalue'}

\item{showCategory}{number of category}

\item{font.size}{font size}

\item{title}{plot title}
}
\value{
plot
}
\description{
dotplot

dotplot for enrichResult
}
\author{
Guangchuang Yu

Guangchuang Yu
}

