% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enrich.internal.R
\name{enrich.internal}
\alias{enrich.internal}
\title{enrich.internal}
\usage{
enrich.internal(gene, organism, pvalueCutoff, pAdjustMethod = "BH", ont,
  universe, minGSSize = 5, qvalueCutoff = 0.2, readable = FALSE, ...)
}
\arguments{
\item{gene}{a vector of entrez gene id.}

\item{organism}{supported organism.}

\item{pvalueCutoff}{Cutoff value of pvalue.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{ont}{Ontology}

\item{universe}{background genes}

\item{minGSSize}{minimal size of genes annotated by Ontology term for testing.}

\item{qvalueCutoff}{cutoff of qvalue}

\item{readable}{whether mapping gene ID to gene Name}

\item{...}{additional parameter}
}
\value{
A \code{enrichResult} instance.
}
\description{
interal method for enrichment analysis
}
\details{
using the hypergeometric model
}
\author{
Guangchuang Yu \url{http://ygc.name}
}
\keyword{manip}

