% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichDO.R
\name{enrichDO}
\alias{enrichDO}
\title{DO Enrichment Analysis}
\usage{
enrichDO(gene, ont = "DO", pvalueCutoff = 0.05, pAdjustMethod = "BH",
  universe, minGSSize = 5, qvalueCutoff = 0.2, readable = FALSE)
}
\arguments{
\item{gene}{a vector of entrez gene id.}

\item{ont}{one of DO or DOLite.}

\item{pvalueCutoff}{Cutoff value of pvalue.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{universe}{background genes}

\item{minGSSize}{minimal size of genes annotated by Ontology term for testing.}

\item{qvalueCutoff}{qvalue Cutoff}

\item{readable}{logical}
}
\value{
A \code{enrichResult} instance.
}
\description{
Given a vector of genes, this function will return the enrichment DO
categories with FDR control.
}
\examples{

data(geneList)
	gene = names(geneList)[geneList > 1]
	yy = enrichDO(gene, pvalueCutoff=0.05)
	summary(yy)

}
\author{
Guangchuang Yu \url{http://guangchuangyu.github.io}
}
\seealso{
\code{\link{enrichResult-class}}
}
\keyword{manip}

