% Generated by roxygen2 (4.0.2): do not edit by hand
\name{enrichMap}
\alias{enrichMap}
\title{enrichMap}
\usage{
enrichMap(x, fixed = TRUE)
}
\arguments{
\item{x}{gseaResult or enrichResult object}

\item{fixed}{if set to FALSE, will invoke tkplot}
}
\value{
figure
}
\description{
enrichment map
}
\details{
enrichment map
}
\author{
G Yu
}

