\docType{class}
\name{enrichResult-class}
\alias{cnetplot,enrichResult-method}
\alias{enrichResult-class}
\alias{plot,enrichResult-method}
\alias{show,enrichResult-method}
\alias{summary,enrichResult-method}
\title{Class "enrichResult"
This class represents the result of enrichment analysis.}
\description{
  Class "enrichResult" This class represents the result of
  enrichment analysis.
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}
\seealso{
  \code{\link{enrichDO}}
}
\keyword{classes}

