% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{enrichResult-class}
\alias{enrichResult-class}
\alias{plot,enrichResult-method}
\alias{show,enrichResult-method}
\alias{summary,enrichResult-method}
\title{Class "enrichResult"
This class represents the result of enrichment analysis.}
\description{
Class "enrichResult"
This class represents the result of enrichment analysis.
}
\section{Slots}{

\describe{
\item{\code{result}}{enrichment analysis}

\item{\code{pvalueCutoff}}{pvalueCutoff}

\item{\code{pAdjustMethod}}{pvalue adjust method}

\item{\code{qvalueCutoff}}{qvalueCutoff}

\item{\code{organism}}{only "human" supported}

\item{\code{ontology}}{biological ontology}

\item{\code{gene}}{Gene IDs}

\item{\code{keytype}}{Gene ID type}

\item{\code{universe}}{background gene}

\item{\code{geneInCategory}}{gene and category association}

\item{\code{gene2Symbol}}{mapping gene to Symbol}

\item{\code{geneSets}}{gene sets}

\item{\code{readable}}{logical flag of gene ID in symbol or not.}
}}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}
\seealso{
\code{\link{enrichDO}}
}
\keyword{classes}

