% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{EXTID2NAME}
\alias{EXTID2NAME}
\title{EXTID2NAME}
\usage{
EXTID2NAME(OrgDb, geneID, keytype)
}
\arguments{
\item{OrgDb}{OrgDb}

\item{geneID}{entrez gene ID}

\item{keytype}{keytype}
}
\value{
gene symbol
}
\description{
mapping gene ID to gene Symbol
}
\author{
Guangchuang Yu \url{http://guangchuangyu.github.io}
}

