% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EXTID2NAME}
\alias{EXTID2NAME}
\title{EXTID2NAME}
\usage{
EXTID2NAME(geneID, organism)
}
\arguments{
\item{geneID}{entrez gene ID}

\item{organism}{one of "human", "mouse" and "yeast"}
}
\value{
gene symbol
}
\description{
mapping gene ID to gene Symbol
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

