% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enrich.internal.R
\name{EXTID2TERMID.DO}
\alias{EXTID2TERMID.DO}
\title{EXTID2TERMID.DO}
\usage{
\method{EXTID2TERMID}{DO}(gene, organism, use.KEGG.db)
}
\arguments{
\item{gene}{gene ID}

\item{organism}{organism}

\item{use.KEGG.db}{logical}
}
\description{
EXTID2TERMID.DO
}

