% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fortify.enrichResult}
\alias{fortify.enrichResult}
\title{fortify}
\usage{
\method{fortify}{enrichResult}(model, data, showCategory = 5, order = FALSE,
  drop = FALSE, ...)
}
\arguments{
\item{model}{enrichResult object}

\item{data}{not use here}

\item{showCategory}{Category numbers to show}

\item{order}{logical}

\item{drop}{logical}

\item{...}{additional parameter}
}
\description{
fortify
}

