% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseAnalyzer.R
\name{fortify.gseaResult}
\alias{fortify.gseaResult}
\title{fortify.gseaResult}
\usage{
\method{fortify}{gseaResult}(model, data, geneSetID, ...)
}
\arguments{
\item{model}{gseaResult object}

\item{data}{not used.}

\item{geneSetID}{gene set ID}

\item{...}{additional parameter}
}
\value{
figure
}
\description{
convert gsea result for ggplot2
}
\author{
G Yu
}

