% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{gene2DO}
\alias{gene2DO}
\title{convert Gene ID to DO Terms}
\usage{
gene2DO(gene)
}
\arguments{
\item{gene}{entrez gene ID}
}
\value{
DO Terms
}
\description{
provide gene ID, this function will convert to the corresponding DO Terms
}
\author{
Guangchuang Yu \url{http://ygc.name}
}

