% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gseAnalyzer.R
\name{getGeneSet.DO}
\alias{getGeneSet.DO}
\title{getGeneSet.DO}
\usage{
\method{getGeneSet}{DO}(setType = "DO", organism)
}
\arguments{
\item{setType}{setType}

\item{organism}{organism}
}
\description{
getGeneSet.DO
}

