% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gseAnalyzer.R
\name{getGeneSet.DOLite}
\alias{getGeneSet.DOLite}
\title{getGeneSet.DOLite}
\usage{
\method{getGeneSet}{DOLite}(setType = "DOLite", organism,
  use_internal_data = TRUE)
}
\arguments{
\item{setType}{setType}

\item{organism}{organism}

\item{use_internal_data}{logical}
}
\description{
getGeneSet.DOLite
}

