% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{GSEA_internal}
\alias{GSEA_internal}
\title{GSEA_internal}
\usage{
GSEA_internal(geneList, geneSets, exponent, nPerm, minGSSize, maxGSSize,
  pvalueCutoff, pAdjustMethod, verbose, seed = FALSE, USER_DATA)
}
\arguments{
\item{geneList}{order ranked geneList}

\item{geneSets}{gene sets}

\item{exponent}{weight of each step}

\item{nPerm}{permutation numbers}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of each geneSet for analyzing}

\item{pvalueCutoff}{p value Cutoff}

\item{pAdjustMethod}{p value adjustment method}

\item{verbose}{print message or not}

\item{seed}{set seed inside the function to make result reproducible. FALSE by default.}

\item{USER_DATA}{annotation data}
}
\value{
gseaResult object
}
\description{
generic function for gene set enrichment analysis
}
\author{
Yu Guangchuang
}

