% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseAnalyzer.R
\name{gseaplot}
\alias{gseaplot}
\title{visualize analyzing result of GSEA}
\usage{
gseaplot(gseaResult, geneSetID, by = "all", title = "")
}
\arguments{
\item{gseaResult}{gseaResult object}

\item{geneSetID}{geneSet ID}

\item{by}{one of "runningScore" or "position"}

\item{title}{plot title}
}
\value{
ggplot2 object
}
\description{
plotting function for gseaResult
}
\author{
Yu Guangchuang
}

