% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,enrichResult,ANY-method}
\alias{plot,gseaResult,ANY-method}
\title{plot method}
\usage{
\S4method{plot}{enrichResult,ANY}(x, type = "bar", ...)

\S4method{plot}{gseaResult,ANY}(x, geneSetID, by = "all", ...)
}
\arguments{
\item{type}{one of bar or cnet}

\item{...}{Additional argument list}

\item{geneSetID}{gent set ID}

\item{by}{one of runningScore, position or all}

\item{x}{the coordinates of points in the plot. Alternatively, a
      single plotting structure, function or \emph{any \R object with a
        \code{plot} method} can be provided.}
}
\value{
plot

plot
}
\description{
plot method generics

plot method for gseaResult
}
\author{
Guangchuang Yu \url{http://ygc.name}

Yu Guangchuang
}

