\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,enrichResult,ANY-method}
\alias{plot,gseaResult,ANY-method}
\title{plot method}
\arguments{
  \item{...}{Additional argument list}

  \item{...}{ignored.}
}
\value{
  plot

  plot
}
\description{
  plot method generics

  plot method for gseaResult
}
\author{
  Guangchuang Yu \url{http://ygc.name}

  Yu Guangchuang
}

