% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enrichDO.R, R/gseAnalyzer.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,enrichResult,ANY-method}
\alias{plot,gseaResult,ANY-method}
\title{plot method}
\usage{
\S4method{plot}{enrichResult,ANY}(x, type = "bar", ...)

\S4method{plot}{gseaResult,ANY}(x, type = "gseaplot", ...)
}
\arguments{
\item{x}{A \code{enrichResult} instance}

\item{type}{one of bar, cnet or enrichMap}

\item{...}{Additional argument list}
}
\value{
plot

plot
}
\description{
plot method generics

plot method for gseaResult
}
\author{
Guangchuang Yu \url{http://ygc.name}

Yu Guangchuang
}

