\docType{methods}
\name{setReadable<-}
\alias{setReadable}
\alias{setReadable<-}
\alias{setReadable<-,enrichResult,ANY-method}
\alias{setReadable-methods}
\title{Methods mapping gene ID to gene symbol for \code{enrichResult} instance}
\arguments{
  \item{x}{A \code{enrichResult} instance.}

  \item{value}{readable flag.}
}
\value{
  A \code{enrichResult} instance.
}
\description{
  setReadable method for \code{enrichResult} instance
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}

