% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setReadable.R
\name{setReadable}
\alias{setReadable}
\title{setReadable}
\usage{
setReadable(x, OrgDb, keyType = "auto")
}
\arguments{
\item{x}{enrichResult Object}

\item{OrgDb}{OrgDb}

\item{keyType}{keyType of gene}
}
\value{
enrichResult Object
}
\description{
mapping geneID to gene Symbol
}
\author{
Yu Guangchuang
}
