% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enrichResult.R
\name{setReadable}
\alias{setReadable}
\title{setReadable}
\usage{
setReadable(x)
}
\arguments{
\item{x}{enrichResult Object}
}
\value{
enrichResult Object
}
\description{
mapping geneID to gene Symbol
}
\author{
Yu Guangchuang
}

