% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show}
\alias{show}
\title{show method}
\usage{
show(object)

show(object)
}
\arguments{
\item{object}{A \code{enrichResult} instance.}
}
\value{
message

message
}
\description{
show method for \code{enrichResult} instance

show method for \code{gseaResult} instance
}
\author{
Guangchuang Yu \url{http://ygc.name}

Guangchuang Yu \url{http://ygc.name}
}

