% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simplot.R
\name{simplot}
\alias{simplot}
\title{simplot}
\usage{
simplot(sim, xlab = "", ylab = "", color.low = "white",
  color.high = "red", labs = TRUE, digits = 2, labs.size = 3,
  font.size = 14, readable = FALSE)
}
\arguments{
\item{sim}{similarity matrix}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{color.low}{color of low value}

\item{color.high}{color of high value}

\item{labs}{logical, add text label or not}

\item{digits}{round digit numbers}

\item{labs.size}{lable size}

\item{font.size}{font size}

\item{readable}{TRUE or FALSE}
}
\value{
ggplot object
}
\description{
plotting similarity matrix
}
\author{
Yu Guangchuang
}

