% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichResult.R, R/gseAnalyzer.R
\docType{methods}
\name{[[,enrichResult,ANY,ANY-method}
\alias{[[,enrichResult,ANY,ANY-method}
\alias{[[,gseaResult,ANY,ANY-method}
\title{[[ method}
\usage{
\S4method{[[}{enrichResult,ANY,ANY}(x, term)

\S4method{[[}{gseaResult,ANY,ANY}(x, term)
}
\arguments{
\item{x}{enrichResult object}

\item{term}{term to query}
}
\description{
accessing gene set

accessing gene set
}

