\docType{methods}
\name{summary}
\alias{summary}
\title{summary method}
\arguments{
  \item{object}{A \code{enrichResult} instance.}

  \item{object}{A \code{gseaResult} instance.}
}
\value{
  A data frame

  A data frame
}
\description{
  summary method for \code{enrichResult} instance

  summary method for \code{gseaResult} instance
}
\author{
  Guangchuang Yu \url{http://ygc.name}

  Guangchuang Yu \url{http://ygc.name}
}

