% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichResult.R, R/gseAnalyzer.R
\docType{methods}
\name{summary}
\alias{summary}
\title{summary method}
\usage{
summary(object, ...)

summary(object, ...)
}
\arguments{
\item{object}{A \code{enrichResult} instance.}

\item{...}{additional parameter}
}
\value{
A data frame

A data frame
}
\description{
summary method for \code{enrichResult} instance

summary method for \code{gseaResult} instance
}
\author{
Guangchuang Yu \url{http://guangchuangyu.github.io}

Guangchuang Yu \url{https://guangchuangyu.github.io}
}

