% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enrich.internal.R
\name{TERM2NAME.DO}
\alias{TERM2NAME.DO}
\title{TERM2NAME.DO}
\usage{
\method{TERM2NAME}{DO}(term, organism, use_internal_data)
}
\arguments{
\item{term}{term id}

\item{organism}{organism}

\item{use_internal_data}{logical}
}
\description{
TERM2NAME.DO
}

