% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enrich.internal.R
\name{TERM2NAME.DOLite}
\alias{TERM2NAME.DOLite}
\title{TERM2NAME.DOLite}
\usage{
\method{TERM2NAME}{DOLite}(term, organism, use.KEGG.db)
}
\arguments{
\item{term}{term ID}

\item{organism}{organism}

\item{use.KEGG.db}{logical}
}
\description{
TERM2NAME.DOLite
}

