% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enrich.internal.R
\name{TERMID2EXTID.DO}
\alias{TERMID2EXTID.DO}
\title{TERMID2EXTID.DO}
\usage{
\method{TERMID2EXTID}{DO}(term, organism)
}
\arguments{
\item{term}{term ID}

\item{organism}{organism}
}
\description{
TERMID2EXTID.DO
}

