% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enrich.internal.R
\name{TERMID2EXTID.DOLite}
\alias{TERMID2EXTID.DOLite}
\title{TERMID2EXTID.DOLite}
\usage{
\method{TERMID2EXTID}{DOLite}(term, organism, ...)
}
\arguments{
\item{term}{term ID}

\item{organism}{organism}

\item{...}{additional parameter}
}
\description{
TERMID2EXTID.DOLite
}

