% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{TERMID2EXTID}
\alias{TERMID2EXTID}
\title{TERMID2EXTID}
\usage{
TERMID2EXTID(term, organism, use_internal_data)
}
\arguments{
\item{term}{term ID vector}

\item{organism}{organism}

\item{use_internal_data}{logical}
}
\description{
Mapping Ontology Term ID to External ID
}

